spiritual-assistant/requirements.txt
transformers>=4.42.0
accelerate>=0.30.0
peft>=0.11.1
bitsandbytes>=0.43.1
datasets>=2.19.0
evaluate>=0.4.2
sentencepiece
protobuf<5

# Optional: torch/torchvision; install a version appropriate for your CUDA/CPU
# Example: pip install torch==2.3.1 torchvision==0.18.1 --index-url https://download.pytorch.org/whl/cu121

streamlit>=1.37.0
langdetect>=1.0.9
python-dotenv>=1.0.1
fastapi>=0.111.0
uvicorn>=0.30.0